update some configs

+++ update.sh
@@ -5,10 +5,21 @@
 # @description:
 
 # update i3wm config
-cp $HOME/.config/i3/config config/i3/config
+if [ -f $HOME/.config/i3/config ]; then
+	cp $HOME/.config/i3/config config/i3/config
+fi
 
 # update kernel config
-cp /usr/src/linux/.config kernel/config
+if [ -f /usr/src/linux/.config ]; then
+	cp /usr/src/linux/.config kernel/config
+fi
 
 # update X resources
-cp $HOME/.Xresources X/Xresources
+if [ -f $HOME/.Xresources ]; then
+	cp $HOME/.Xresources X/Xresources
+fi
+
+# update vim config
+if [ -f /etc/vim/vimrc.local ]; then
+	cp /etc/vim/vimrc.local vim/vimrc.local
+fi
